pyproject.toml
[tool.poetry]
name = "sipag-api"
version = "0.1.0"
description = ""
authors = ["Jhonas Palad <jhonasemmanuel@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
psycopg = "^3.2.1"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
blurhash-python = "^1.2.2"
djangorestframework-gis = "^1.0"
channels = {extras = ["daphne"], version = "^4.1.0"}
python-dotenv = "^1.0.1"
django-storages = {extras = ["s3"], version = "^1.14.4"}
channels-redis = "^4.2.0"
coverage = "^7.6.1"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
